move Voila loader to standalone loader
